requirements.txt
autopep8==1.5.7
bitarray==2.2.2
cycler==0.10.0
jellyfish==0.8.2
kiwisolver==1.3.1
matplotlib==3.4.2
mypy-extensions==0.4.3
numpy==1.21.1
Pillow==8.3.1
prettytable==2.1.0
prompt-toolkit==3.0.19
pycodestyle==2.7.0
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-nubia==0.2b5
six==1.16.0
termcolor==1.1.0
toml==0.10.2
typing-extensions==3.10.0.0
typing-inspect==0.7.1
wcwidth==0.2.5
